requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
pytest
pytest-flask